pyproject.toml
[project]
name = "music-app"
version = "0.1.0"
description = ""
authors = [
    {name = "tony",email = "technicalprogram4658@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.113.0",
    "sqlalchemy[asyncio]>=2.0.41,<3.0.0",
    "uvicorn[standard]>=0.34.3,<0.35.0",
    "pydantic-settings>=2.9.1,<3.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "alembic>=1.16.1,<2.0.0",
    "black>=25.1.0,<26.0.0",
    "greenlet>=3.2.3,<4.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "bcrypt==4.0.1",
    "pydantic[email]>=2.11.7,<3.0.0",
    "python-multipart>=0.0.20,<0.0.21",
    "python-jose>=3.5.0,<4.0.0",
    "redis[hiredis]>=6.2.0,<7.0.0",
    "yt-dlp[nightly]>=2025.6.30,<2026.0.0",
    "pip-system-certs>=5.2,<6.0",
    "aioboto3>=14.3.0,<15.0.0",
    "types-aiobotocore-s3>=2.23.0,<3.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.9",
]
test = [
    "pytest>=8.4.0",
    "pytest-asyncio>=0.24.0",
]

[tool.black]
line-length = 88
